+++ Changelog.md
@@ -1,7 +1,15 @@
 # Vanilo Euplatesc Module Changelog
 
+## Unreleased
+##### 2021-XX-YY
+
+- Renamed package from `konekt/vanilo-euplatesc` to `vanilo/euplatesc`
+- Changed the namespace from `Konekt\Euplatesc` to `Vanilo\Euplatesc`
+- Requires Vanilo 2.2+
+- Dropped PHP 7.3 support
+
 ## 1.0.0
 ##### 2021-01-05
 
-- Initial release
+- Initial release as `konekt/vanilo-euplatesc`
 - Proven to work with PHP 7.4 - 8.0, Laravel 6 - 8